Gameplay Changes
Changed values for player and boss to balance gameplay. Recommend fixing boss stun, as the boss often is not stunned when hitting swapped objects. Also, recommend adding walls to keep the boss from jumping off the level. And lastly, please add a way for the rocks to respawn once they are destroyed.